Phase 3: migrate a whole database between backends
main.py --cli migrate-db --target mysql://user@host:3306/db, plus
core/db_migrate.py. Copies table by table in foreign-key order,
preserving primary keys so selections and overrides stay attached, and
reads through raw SQL so values move untranslated. Batched at 500 rows
(~0.3 MB against a 16 MB max_allowed_packet). The target password comes
from $CM_TARGET_DB_PASSWORD rather than the URL, which is visible in the
process list. A non-empty target is refused unless --force.

Not a JSON export/import round trip: that path drops every
TrackAnalysis, the file mtime/size incremental scanning needs, and ten
other Track columns.

Each table is verified after writing by re-reading it and comparing a
content hash of normalized values — and that immediately caught a bug
that row counts could not.

Track.file_mtime was a FloatField, which peewee maps to MySQL FLOAT:
single precision, ~7 significant digits. A Unix timestamp needs 17, so
1666807963.287016 came back as 1666810000.0 — over half an hour out.
Every file would have looked modified to an incremental scan, and
analysis restore compares on exactly this value, so all 6,373 analyses
would have been silently recomputed. Every row count was correct; only
the content hash reported it. Every FloatField is now a DoubleField
(Track.file_mtime, TrackAnalysis.volatility, AnalysisSnapshot.volatility
and .file_mtime). SQLite has one numeric type and is unaffected, which
is precisely why this could not surface without a server.

The other two mismatches were representation rather than loss: SQLite
keeps tz-aware timestamps as text carrying +00:00, while MySQL DATETIME
has no timezone and stores whole seconds. Normalization now compares
instants, so those verify equal; MySQL holds naive UTC.

Real data: a copy of the production library — 22,701 rows, 7,279 tracks,
5,420 works, 6,373 analyses — migrates in ~2.5 s, every table verified.
generate-all against MariaDB produced 22 of 24 playlists with track sets
identical to SQLite. The other two also differ between two SQLite runs,
so that is the shuffle on length-limited profiles, not the backend.

Tests: 226 green on SQLite (12 new for migration, including one that
corrupts the read-back to prove verification fails when it should), and
8 against MariaDB 10.6 including a cross-backend round trip pinning the
mtime precision.

+++ music_manager/core/database.py
@@ -176,7 +176,7 @@ def initialize_database(db_path: Path | None = None,
         logger.info("Migrated: added work_tag, mb_work_id to tracks")
     if "file_mtime" not in track_cols:
         run_migrate(
-            migrator.add_column("tracks", "file_mtime", pw.FloatField(null=True)),
+            migrator.add_column("tracks", "file_mtime", pw.DoubleField(null=True)),
             migrator.add_column("tracks", "file_size", pw.IntegerField(null=True)),
         )
         logger.info("Migrated: added file_mtime, file_size to tracks")
@@ -354,7 +354,11 @@ class Track(BaseModel):
     ensemble = pw.TextField(null=True)           # orchestra/ensemble
     work_tag = pw.TextField(null=True)          # raw WORK tag from file
     mb_work_id = pw.TextField(null=True)        # per-track MusicBrainz work ID from file
-    file_mtime = pw.FloatField(null=True)       # file modification time (os.stat)
+    # DOUBLE, not FLOAT: a Unix timestamp needs ~17 significant digits and
+    # MySQL's single-precision FLOAT keeps ~7, which shifted real mtimes by
+    # over half an hour — every file would look changed to an incremental
+    # scan, and analysis preservation compares on this value.
+    file_mtime = pw.DoubleField(null=True)      # file modification time (os.stat)
     file_size = pw.IntegerField(null=True)       # file size in bytes
     # When this track first entered the library. Set on INSERT only and
     # preserved across full rescans — file_mtime is the FILE's time and
